pyproject.toml
[project]
name = "gemini-video-analyze-mcp"
version = "0.1.0"
description = "MCP server for video analysis using Gemini API"
requires-python = ">=3.12"
scripts = { "gemini-video-analyze-mcp" = "src.server:main" }
dependencies = [
    "google-genai>=1.11.0",
    "mcp[cli]>=1.19.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.14.0",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
